host/pyproject.toml
[project]
name = "midihelper"
version = "0.1.0"
description = "Host computer program of the Midi Choir Project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mido>=1.3.3",
    "pyserial>=3.5",
]
